Created CardLink component
