+++ src/web/components/Card/Card.tsx
@@ -8,40 +8,10 @@ import { ImageLeftLayout } from './ImageLeftLayout';
 import { ImageRightLayout } from './ImageRightLayout';
 import { ImageWrapper } from './ImageWrapper';
 import { TopBar } from './TopBar';
+import { CardLink } from './CardLink';
 
 import { SmallHeadline } from '@frontend/web/components/SmallHeadline';
 import { ImageComponent } from '@frontend/web/components/elements/ImageComponent';
-
-const linkStyles = ({
-    backgroundColour,
-    backgroundOnHover,
-}: {
-    backgroundColour: string;
-    backgroundOnHover: string;
-}) => css`
-    /* a tag specific styles */
-    color: inherit;
-    text-decoration: none;
-
-    /* The whole card is one link so we card level styles here */
-    width: 100%;
-    background-color: ${backgroundColour};
-    :hover {
-        background-color: ${backgroundOnHover};
-    }
-
-    /* Sometimes a headline contains it's own link so we use the
-       approach described below to deal with nested links
-       See: https://css-tricks.com/nested-links/ */
-    :before {
-        content: '';
-        position: absolute;
-        left: 0;
-        top: 0;
-        right: 0;
-        bottom: 0;
-    }
-`;
 
 const listStyles = css`
     /* Here we ensure the card stretches to fill the containing space */
@@ -101,12 +71,10 @@ export const Card = ({
     const Layout = decideLayout(image);
     return (
         <li className={listStyles}>
-            <a
-                href={linkTo}
-                className={linkStyles({
-                    backgroundColour: palette.neutral[93],
-                    backgroundOnHover: palette.neutral[86],
-                })}
+            <CardLink
+                linkTo={linkTo}
+                backgroundColour={palette.neutral[93]}
+                backgroundOnHover={palette.neutral[86]}
             >
                 <TopBar topBarColour={palette[pillar].main}>
                     <Layout>
@@ -130,7 +98,7 @@ export const Card = ({
                         </>
                     </Layout>
                 </TopBar>
-            </a>
+            </CardLink>
         </li>
     );
 };
